chore: moved street_details_sheet to a new folder
